requirements.txt
gTTS
pygame
Pillow
cvzone
numpy
opencv-python
pytesseract
Pillow
matplotlib
